Move callers to their own header
